Flamegraph shows that lots of work is going for cleaning the tensors using NDManager, need to find an alternative solution otherwise delegate to Cleaner PtNDArray

+++ src/lib/ai/djl/pytorch/engine/PtNDArray.java
@@ -62,7 +62,7 @@ public class PtNDArray extends NativeResource<Long> implements NDArray {
     // use Boolean object to maintain three status: null, false, true
     private Boolean hasGradient;
     private PtNDManager manager;
-    private byte taskPossession = 0;
+    private int taskPossession = 0;
     // keep a reference to direct buffer to avoid GC release the memory
     @SuppressWarnings("PMD.UnusedPrivateField")
     private ByteBuffer[] dataRef;
@@ -1764,7 +1764,7 @@ public class PtNDArray extends NativeResource<Long> implements NDArray {
     }
 
     public void setTaskPossession(int taskPossession) {
-        this.taskPossession = (byte) taskPossession;
+        this.taskPossession = taskPossession;
     }
 
     /**
